feat: separate melBands into its own extractor
